improved partlist generation by making it dynamic with the products

+++ src/main/java/FunctionLayer/calculators/RoofCalculator.java
@@ -12,7 +12,7 @@ package FunctionLayer.calculators;
 public class RoofCalculator {
     
     public double RoofCalc (double a, double b){
-        double rc = (int) (a * b);
+        double rc = a * b;
         return rc;
     }
